Fix the treatment of broken symlink
`os.path.exists(f)` and `os.path.isfile(f)` return `False` when the `f`
is a broken symlink. This causes dotty to skip the entire existing-file
logic and try to link an existing file, resulting in `FileExistsError:
[Errno 17]`. Even if the logic was used, dotty attempts to remove the
broken symlink as directory, because `os.path.isfile(f)` says it's not
a file.

Using `os.path.lexists` and determining whether the symlink is broken
fixes the problem.